Change triangle vertex colors to {R,G,B} and used interpolated values.
This extends the earlier example to give the triangle Red, Green, and Blue
vertices rather than everything being uniformly Android green.  The colors
between vertices are interpolated and used in the fragment shader, leading to a
colorful triangle.  This change was based off a bit of trial and error with
`in`, `out`, `attribute`, and `varying` types in OpenGL until it worked.  The
final code in this commit is closest to that found in
https://arm-software.github.io/opengl-es-sdk-for-android/introduction_to_shaders.html
.